refactor(mcp): split instances.rs into an instances/ module
Pure file move, no behavior change. The 2.3k-line instances.rs becomes:

  instances/mod.rs         shared helpers: PropView, find_property,
                           resolve_class, fetch_instance, URI minting
  instances/validate.rs    SHACL-derived property validation
  instances/describe.rs    describe_perspective + schema projection
  instances/create.rs      instance_create
  instances/query.rs       instance_query + instance_get
  instances/update.rs      instance_update + instance_remove
  instances/collections.rs instance_add_to_collection
  instances/tests.rs       the existing integration tests, unchanged

Only visibility changed: helpers the submodules call went from private to
pub(super) (describe_value to pub(crate) so mod.rs can re-export it), and
the parameter structs are re-exported from mod.rs so external users
(harness_bridge, tests) keep their `instances::*` paths.
